1) Move oss-health to use Nomad logic 2) Create a factory for project managers
